+++ block/src/factory.rs
@@ -21,8 +21,8 @@ use crate::block_io_uring_is_supported;
 use crate::disk_file::AsyncFullDiskFile;
 use crate::error::{BlockError, BlockErrorKind, BlockResult};
 use crate::fixed_vhd_disk::VhdDisk;
+use crate::formats::raw::{RawBackend, RawDisk};
 use crate::qcow_disk::QcowDisk;
-use crate::raw_disk::{RawBackend, RawDisk};
 use crate::vhdx_sync::VhdxDisk;
 use crate::{
     ImageType, block_aio_is_supported, detect_image_type, open_disk_image, preallocate_disk,
