+++ block/src/fixed_vhd_sync.rs
@@ -9,7 +9,7 @@ use std::os::unix::io::RawFd;
 use vmm_sys_util::eventfd::EventFd;
 
 use crate::async_io::{AsyncIo, AsyncIoCompletion, AsyncIoError, AsyncIoOperation, AsyncIoResult};
-use crate::raw_sync::RawSync;
+use crate::formats::raw::worker::sync::RawSync;
 
 pub struct FixedVhdSync {
     raw_file_sync: RawSync,
